Move to pkg/logging.
